chat-ui/requirements.txt
streamlit
google-cloud-aiplatform>=1.71.0
requests
protobuf
google-auth
